requirement.txt
pandas
tabula-py
csv
re
PyPDF2